Patch date range formatter to render dates
Currently the module will output the formatted range directly as
markup. We want to be able to modify the output so having a render
array for this or the like is nice.

This merge request changes the output of the formatter to a render
array using the time #type. This is also nice for semantics.

The merge request has already been approved and merged so we can drop
the patch once a new version of the module has been released.